make OrderBook even dumber
